+++ src/Rfm69OcfDeviceController.h
@@ -13,7 +13,7 @@ signals:
 
 public slots:
 private:
-    uint16_t readPacket(uint8_t* buf, uint16_t maxSize, String* address) override;
+    int readPacket(uint8_t* buf, uint16_t maxSize, String* address) override;
     bool init() override;
     void send_packet(COAPPacket* packet) override;
